# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title": "Briefed API", "version": "1.1.0", "description": "Public and optionally personalized RSS and Atom timeline API."},
  "servers": [{"url": "http://localhost:8080"}],
...
  "components": {
    "securitySchemes": {"sessionCookie": {"type":"apiKey","in":"cookie","name":"briefed_session","description":"HTTP-only cookie set by register/login."}},
    "schemas": {
      "Error": {"type":"object","required":["error"],"properties":{"error":{"type":"string"}}},
      "UUID": {"type":"string","format":"uuid"},
...
      "User": {"type":"object","properties":{"id":{"$ref":"#/components/schemas/UUID"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"},"name":{"type":"string"},"email":{"type":"string","format":"email"}}},
...
      "RegisterRequest": {"type":"object","required":["name","email","password"],"properties":{"name":{"type":"string"},"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password","minLength":8},"category_ids":{"type":"array","items":{"$ref":"#/components/schemas/UUID"}},"feed_ids":{"type":"array","items":{"$ref":"#/components/schemas/UUID"}}}},
      "LoginRequest": {"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}}},
...
    },
    "responses": {
      "BadRequest": {"description":"Invalid request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},
      "Unauthorized": {"description":"Authentication required","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}
    }
  },
  "paths": {
...
    "/v1/auth/register": {"post":{"tags":["Authentication"],"summary":"Register and start a session","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"},"example":{"name":"Ada Lovelace","email":"ada@example.com","password":"password123","category_ids":[],"feed_ids":[]}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"409":{"description":"Email already exists"}}}},
    "/v1/auth/login": {"post":{"tags":["Authentication"],"summary":"Login and start a session","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"},"example":{"email":"ada@example.com","password":"password123"}}}},"responses":{"200":{"description":"Authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"}}}},
    "/v1/auth/logout": {"post":{"tags":["Authentication"],"summary":"End the current session","security":[{"sessionCookie":[]}],"responses":{"204":{"description":"Logged out"}}}},
...
  }
}